+++ lib_src/api/asynchronous_fifo.h
@@ -59,7 +59,7 @@ typedef struct asynchronous_fifo_t_ {
  * The ``state`` argument should be an int64_t array of
  * ``ASYNCHRONOUS_FIFO_INT64_ELEMENTS`` elements that is cast to
  * ``asynchronous_fifo_t*``.
- * 
+ *
  * That pointer should also be used for all other operations, including operations
  * both the consumer and producer sides.
  *
@@ -177,8 +177,6 @@ void asynchronous_fifo_exit(asynchronous_fifo_t * UNSAFE state);
  *
  * @param   timestamp           The number of ticks when this sample was input.
  *
- * @param   xscope_used         Set to 1 if the PID values should be output over
- *                              xscope. Used for debugging. This parameter is subject to be removed in future revisions.
  *
  * @returns The current estimate of the mismatch of input and output frequencies.
  *          This is represented as a 32-bit signed number. Zero means no mismatch,
@@ -193,8 +191,7 @@ void asynchronous_fifo_exit(asynchronous_fifo_t * UNSAFE state);
 int32_t asynchronous_fifo_producer_put(asynchronous_fifo_t * UNSAFE state,
                                        int32_t * UNSAFE samples,
                                        int n,
-                                       int32_t timestamp,
-                                       int xscope_used);
+                                       int32_t timestamp);
 
 
 /**