+++ tests/asynchronous_fifo_asrc_test/src/asynchronous_fifo_asrc_test.c
@@ -92,9 +92,9 @@ int base_frequency_multiplier(int frequency) {
 
 extern int async_resets;
 
-DECLARE_JOB(producer,   (asynchronous_fifo_t *, int, int, int, int *));
-DECLARE_JOB(consumer,   (asynchronous_fifo_t *, int, int));
-DECLARE_JOB(test_async, (int, int, int, int *));
+DECLARE_JOB(producer,   (asynchronous_fifo_t *, int, int, int *));
+DECLARE_JOB(consumer,   (asynchronous_fifo_t *, int));
+DECLARE_JOB(test_async, (int, int, int *));
 
 #define seconds 10
 #define OFFSET 0 // 0x70000000
@@ -160,7 +160,7 @@ static int interpolation_ticks_2D[6][6] = {
     {  2083, 2083, 1042, 1042,  521,  521}
 };
 
-void producer(asynchronous_fifo_t *a, int input_frequency, int output_frequency, int xscope_used, int *errors) {
+void producer(asynchronous_fifo_t *a, int input_frequency, int output_frequency, int *errors) {
     int interpolation_ticks = interpolation_ticks_2D[fs_code(input_frequency)][fs_code(output_frequency)];
     asrc_state_t sASRCState[SRC_CHANNELS_PER_INSTANCE];                                   // ASRC state machine state
     int iASRCStack[SRC_CHANNELS_PER_INSTANCE][ASRC_STACK_LENGTH_MULT * SRC_N_IN_SAMPLES * 100]; // Buffer between filter stages
@@ -225,9 +225,10 @@ void producer(asynchronous_fifo_t *a, int input_frequency, int output_frequency,
         if (num_samples) {
             asm volatile("gettime %0" : "=r" (t2));
             int ts = asrc_timestamp_interpolation(now, &sASRCCtrl[0], interpolation_ticks);
-            if (xscope_used) xscope_int(5, fs_ratio >> 32);
-            error = asynchronous_fifo_producer_put(a, (int32_t *)out_samples, num_samples, ts+OFFSET,
-                                              xscope_used);
+#if defined(ASYNC_FIFO_XSCOPE_INSTRUMENTATION)
+            xscope_int(5, fs_ratio >> 32);
+#endif
+            error = asynchronous_fifo_producer_put(a, (int32_t *)out_samples, num_samples, ts+OFFSET);
             asm volatile("gettime %0" : "=r" (t3));
             if (i == 48008) {
 //                printf("%d %d %d %d\n", t1-t0, t2-t1, t3-t2, t3-t0);
@@ -255,7 +256,7 @@ void producer(asynchronous_fifo_t *a, int input_frequency, int output_frequency,
     }
 }
 
-void consumer(asynchronous_fifo_t *a, int output_frequency, int xscope_used) {
+void consumer(asynchronous_fifo_t *a, int output_frequency) {
     hwtimer_t tmr = hwtimer_alloc();
     uint64_t now = hwtimer_get_time(tmr);
     int freq = POSITIVE_DEVIATION(output_frequency);
@@ -274,7 +275,9 @@ void consumer(asynchronous_fifo_t *a, int output_frequency, int xscope_used) {
         hwtimer_set_trigger_time(tmr, now);
         (void) hwtimer_get_time(tmr);
         asynchronous_fifo_consumer_get(a, &output_data, now + OFFSET);
-        if (xscope_used) xscope_int(0, output_data);
+#if defined(ASYNC_FIFO_XSCOPE_INSTRUMENTATION)
+        xscope_int(0, output_data);
+#endif
         if (i == output_frequency/2) {
             freq = NEGATIVE_DEVIATION(output_frequency);
             step = 100000000 / freq;
@@ -286,8 +289,7 @@ void consumer(asynchronous_fifo_t *a, int output_frequency, int xscope_used) {
 
 #define FIFO_LENGTH   100
 
-void test_async(int input_frequency, int output_frequency, int xscope_used,
-                int *errors) {
+void test_async(int input_frequency, int output_frequency, int *errors) {
     int64_t array[ASYNCHRONOUS_FIFO_INT64_ELEMENTS(FIFO_LENGTH, 1)];
     asynchronous_fifo_t *asynchronous_fifo_state = (asynchronous_fifo_t *)array;
 
@@ -298,8 +300,8 @@ void test_async(int input_frequency, int output_frequency, int xscope_used,
                                         fs_code(output_frequency));
 
     PAR_JOBS(
-        PJOB(producer, (asynchronous_fifo_state, input_frequency, output_frequency, xscope_used, errors)),
-        PJOB(consumer, (asynchronous_fifo_state, output_frequency, xscope_used))
+        PJOB(producer, (asynchronous_fifo_state, input_frequency, output_frequency, errors)),
+        PJOB(consumer, (asynchronous_fifo_state, output_frequency))
         );
     asynchronous_fifo_exit(asynchronous_fifo_state);
 
@@ -309,10 +311,10 @@ int test_44100_low() {
     int e0=0, e1=0, e2=0, e3=0;
     printf("Testing 44100 low\n");
     PAR_JOBS(
-        PJOB(test_async, (44100, 44100, 0, &e0)), // OK, 4 sec or so.
-        PJOB(test_async, (44100, 48000, 0, &e1)), // OK, 4 sec or so. Slight overshoot
-        PJOB(test_async, (44100, 88200, 0, &e2)), // OK, 4 sec or so.
-        PJOB(test_async, (44100, 96000, 0, &e3)) // OK  4 sec or so.
+        PJOB(test_async, (44100, 44100, &e0)), // OK, 4 sec or so.
+        PJOB(test_async, (44100, 48000, &e1)), // OK, 4 sec or so. Slight overshoot
+        PJOB(test_async, (44100, 88200, &e2)), // OK, 4 sec or so.
+        PJOB(test_async, (44100, 96000, &e3)) // OK  4 sec or so.
         );
     return e0 + e1 + e2 + e3;
 }
@@ -321,10 +323,10 @@ int test_48000_low() {
     int e0=0, e1=0, e2=0, e3=0;
     printf("Testing 48000 low\n");
     PAR_JOBS(
-        PJOB(test_async, (48000, 44100, 0, &e0)), // OK, 4 sec or so
-        PJOB(test_async, (48000, 48000, 0, &e1)), // OK  4.16 s stable slight overshoot
-        PJOB(test_async, (48000, 88200, 0, &e2)), // OK, 4 sec or so
-        PJOB(test_async, (48000, 96000, 0, &e3)) // OK  4.16 s stable
+        PJOB(test_async, (48000, 44100, &e0)), // OK, 4 sec or so
+        PJOB(test_async, (48000, 48000, &e1)), // OK  4.16 s stable slight overshoot
+        PJOB(test_async, (48000, 88200, &e2)), // OK, 4 sec or so
+        PJOB(test_async, (48000, 96000, &e3)) // OK  4.16 s stable
         );
     return e0 + e1 + e2 + e3;
 }
@@ -333,10 +335,10 @@ int test_4xx00_high() {
     int e0, e1, e2, e3;
     printf("Testing 44100/48000 high\n");
     PAR_JOBS(
-        PJOB(test_async, (44100, 176400, 0, &e2)), // OK, 4 sec or so.
-        PJOB(test_async, (44100, 192000, 0, &e3)), // OK  4 sec or so.
-        PJOB(test_async, (48000, 176400, 0, &e0)), // OK, 4 sec or so
-        PJOB(test_async, (48000, 192000, 0, &e1)) // OK  4.16 s stable
+        PJOB(test_async, (44100, 176400, &e2)), // OK, 4 sec or so.
+        PJOB(test_async, (44100, 192000, &e3)), // OK  4 sec or so.
+        PJOB(test_async, (48000, 176400, &e0)), // OK, 4 sec or so
+        PJOB(test_async, (48000, 192000, &e1)) // OK  4.16 s stable
         );
     return e0 + e1 + e2 + e3;
 }
@@ -345,10 +347,10 @@ int test_88200_low() {
     int e0, e1, e2, e3;
     printf("Testing 88200 low\n");
     PAR_JOBS(
-        PJOB(test_async, (88200, 44100, 0, &e0)), // OK, 4 sec or so
-        PJOB(test_async, (88200, 48000, 0, &e1)), // OK, 4 sec or so. Slight overshoot, phase error not zero at the end?
-        PJOB(test_async, (88200, 88200, 0, &e2)), // OK, 4 sec or so. Slight overshoot
-        PJOB(test_async, (88200, 96000, 0, &e3))  // OK  4 sec or so. Slight overshoot
+        PJOB(test_async, (88200, 44100, &e0)), // OK, 4 sec or so
+        PJOB(test_async, (88200, 48000, &e1)), // OK, 4 sec or so. Slight overshoot, phase error not zero at the end?
+        PJOB(test_async, (88200, 88200, &e2)), // OK, 4 sec or so. Slight overshoot
+        PJOB(test_async, (88200, 96000, &e3))  // OK  4 sec or so. Slight overshoot
         );
     return e0 + e1 + e2 + e3;
 }
@@ -357,10 +359,10 @@ int test_96000_low() {
     int e0=0, e1=0, e2=0, e3=0;
     printf("Testing 96000 low\n");
     PAR_JOBS(
-        PJOB(test_async, (96000, 44100, 0, &e0)), // OK, 4 sec or so. Non zero phase difference
-        PJOB(test_async, (96000, 48000, 0, &e1)), // OK  4.16 s stable slight overshoot
-        PJOB(test_async, (96000, 88200, 0, &e2)), // OK, 4 sec or so. Phase error not zero
-        PJOB(test_async, (96000, 96000, 0, &e3)) // OK  4.14 s stable
+        PJOB(test_async, (96000, 44100, &e0)), // OK, 4 sec or so. Non zero phase difference
+        PJOB(test_async, (96000, 48000, &e1)), // OK  4.16 s stable slight overshoot
+        PJOB(test_async, (96000, 88200, &e2)), // OK, 4 sec or so. Phase error not zero
+        PJOB(test_async, (96000, 96000, &e3)) // OK  4.14 s stable
         );
     return e0 + e1 + e2 + e3;
 }
@@ -369,10 +371,10 @@ int test_9xx00_high() {
     int e0=0, e1=0, e2=0, e3=0;
     printf("Testing 44100/48000 high\n");
     PAR_JOBS(
-        PJOB(test_async, (88200, 176400, 0, &e2)), // OK, 4 sec or so. Slight overshoot
-        PJOB(test_async, (88200, 192000, 0, &e3)), // OK  4 sec or so. Slight overshoot
-        PJOB(test_async, (96000, 176400, 0, &e0)), // OK, 4 sec or so. Phase error not zero
-        PJOB(test_async, (96000, 192000, 0, &e1)) // OK  4.14 s stable
+        PJOB(test_async, (88200, 176400, &e2)), // OK, 4 sec or so. Slight overshoot
+        PJOB(test_async, (88200, 192000, &e3)), // OK  4 sec or so. Slight overshoot
+        PJOB(test_async, (96000, 176400, &e0)), // OK, 4 sec or so. Phase error not zero
+        PJOB(test_async, (96000, 192000, &e1)) // OK  4.14 s stable
         );
     return e0 + e1 + e2 + e3;
 }
@@ -381,10 +383,10 @@ int test_176400_low() {
     int e0=0, e1=0, e2=0, e3=0;
     printf("Testing 176400 low\n");
     PAR_JOBS(
-        PJOB(test_async, (176400, 44100, 0, &e0)),
-        PJOB(test_async, (176400, 48000, 0, &e1)),
-        PJOB(test_async, (176400, 88200, 0, &e2)),
-        PJOB(test_async, (176400, 96000, 0, &e3)) // phase != 0?
+        PJOB(test_async, (176400, 44100, &e0)),
+        PJOB(test_async, (176400, 48000, &e1)),
+        PJOB(test_async, (176400, 88200, &e2)),
+        PJOB(test_async, (176400, 96000, &e3)) // phase != 0?
         );
     return e0 + e1 + e2 + e3;
 }
@@ -393,10 +395,10 @@ int test_192000_low() {
     int e0=0, e1=0, e2=0, e3=0;
     printf("Testing 192000 low\n");
     PAR_JOBS(
-        PJOB(test_async, (192000, 44100, 0, &e0)),
-        PJOB(test_async, (192000, 48000, 0, &e1)),
-        PJOB(test_async, (192000, 88200, 0, &e2)), 
-        PJOB(test_async, (192000, 96000, 0, &e3))
+        PJOB(test_async, (192000, 44100, &e0)),
+        PJOB(test_async, (192000, 48000, &e1)),
+        PJOB(test_async, (192000, 88200, &e2)),
+        PJOB(test_async, (192000, 96000, &e3))
         );
     return e0 + e1 + e2 + e3;
 }
@@ -405,10 +407,10 @@ int test_1xxx00_high() {
     int e0=0, e1=0, e2=0, e3=0;
     printf("Testing 176400/192000 high\n");
     PAR_JOBS(
-        PJOB(test_async, (176400, 176400, 0, &e2)),
-        PJOB(test_async, (176400, 192000, 0, &e3)),
-        PJOB(test_async, (192000, 176400, 0, &e0)),
-        PJOB(test_async, (192000, 192000, 0, &e1))
+        PJOB(test_async, (176400, 176400, &e2)),
+        PJOB(test_async, (176400, 192000, &e3)),
+        PJOB(test_async, (192000, 176400, &e0)),
+        PJOB(test_async, (192000, 192000, &e1))
         );
     return e0 + e1 + e2 + e3;
 }
@@ -424,7 +426,7 @@ int main(void) {
     errors += test_9xx00_high();
     errors += test_176400_low();
     errors += test_192000_low();
-    errors += test_1xxx00_high(); 
+    errors += test_1xxx00_high();
     if (errors == 0) {
         printf("PASS\n");
     } else {