+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 lib_src change log
 ==================
 
+3.0.0
+-----
+
+  * FIXED: Double buffer asrc_io.input_timestamp to prevent producer timestamp
+    getting overwritten during asrc processing
+  * REMOVED: xscope_used argument from the asynchronous_fifo_producer_put() API
+
 2.5.0
 -----